Renamed files and functions
file scs_options.py to param_wrappers.py
click_options to add_params
