+++ exasol/nb_connector/cli/options/saas.py
@@ -1,5 +1,5 @@
 from exasol.nb_connector.ai_lab_config import AILabConfig as CKey
-from exasol.nb_connector.cli.scs_options import (
+from exasol.nb_connector.cli.param_wrappers import (
     ScsOption,
     ScsSecretOption,
 )
