+++ exasol/nb_connector/cli/options/ssl.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 from exasol.nb_connector.ai_lab_config import AILabConfig as CKey
-from exasol.nb_connector.cli.scs_options import ScsOption
+from exasol.nb_connector.cli.param_wrappers import ScsOption
 
 SSL_OPTIONS = [
     ScsOption(
